Update to imglib2 5.8.0
And add backwards incompatibilities in behavior.

See also:
https://github.com/imglib/imglib2/commit/1c105ccbb735fea9b72337c8b0f7e4021f3141fc

+++ src/main/java/net/imagej/ops/convert/Types.java
@@ -183,7 +183,7 @@ public final class Types {
 
 	public static float float32(final double value) { return (float) value; }
 
-	public static float float32(final BigInteger value) { return value.and(mask128).longValue(); }
+	public static float float32(final BigInteger value) { return value.and(mask128).floatValue(); }
 
 	public static float float32(final boolean value) { return value ? 1 : 0; }
 
@@ -191,7 +191,7 @@ public final class Types {
 
 	public static double float64(final double value) { return value; }
 
-	public static double float64(final BigInteger value) { return value.and(mask128).longValue(); }
+	public static double float64(final BigInteger value) { return value.and(mask128).doubleValue(); }
 
 	public static double float64(final boolean value) { return value ? 1 : 0; }
